Move map constants to constants file
